# docs/openapi-authoring.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetryRequest": {
        "properties": {
          "from_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 노드부터 재실행 힌트(현재는 전체 재실행으로 폴백).",
            "title": "From Node"
          }
        },
        "title": "RetryRequest",
        "type": "object"
      },
      "RunBulkDeleteRequest": {
        "properties": {
          "ids": {
            "description": "삭제할 run_id 목록",
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "title": "RunBulkDeleteRequest",
        "type": "object"
      },
      "RunDetail": {
        "description": "run 상세 — 노드별 상태 스냅샷 포함.",
        "properties": {
          "by_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By User"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "failed_at_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At Node"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "node_states": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Node States",
            "type": "object"
          },
          "problem_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Id"
          },
          "problem_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Title"
          },
          "saved_count": {
            "default": 0,
            "title": "Saved Count",
            "type": "integer"
          },
          "saved_problem_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Saved Problem Ids",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "default": "running",
            "title": "Status",
            "type": "string"
          },
          "target_count": {
            "default": 0,
            "title": "Target Count",
            "type": "integer"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "id"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunRequest": {
        "examples": [
          {
            "count": 5,
            "problem_id": 1
          }
        ],
        "properties": {
          "by_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "실행한 관리자 표시명 (RunsView 메타용, 옵션)",
            "title": "By User"
          },
          "count": {
            "default": 5,
            "description": "생성할 변형 개수 (검증/심사 통과 시에만 저장됨)",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "problem_id": {
            "description": "변형의 부모가 될 원본 문제 ID",
            "examples": [
              1
            ],
            "title": "Problem Id",
            "type": "integer"
          }
        },
        "required": [
          "problem_id"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "run_id": {
            "description": "이 실행에 대한 메모리 큐 키 — SSE에서 사용",
            "title": "Run Id",
            "type": "string"
          },
          "trace_id": {
            "description": "LangSmith 트레이스 ID (UUID). /api/spans/{trace_id}로 조회.",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "trace_id"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunSummary": {
        "description": "run 목록 행 — 무거운 node_states 제외.",
        "properties": {
          "by_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By User"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "failed_at_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At Node"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "problem_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Id"
          },
          "problem_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Title"
          },
          "saved_count": {
            "default": 0,
            "title": "Saved Count",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "default": "running",
            "title": "Status",
            "type": "string"
          },
          "target_count": {
            "default": 0,
            "title": "Target Count",
            "type": "integer"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "id"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "LangGraph 파이프라인으로 원본 문제로부터 변형을 생성하는 출제 엔진. DB·sandbox는 backend·judge_engine에 위임 — 내부 HTTP API로 통신. /api/health를 제외한 라우트는 Bearer admin 토큰을 요구한다.",
    "title": "JCodeQuest Authoring Engine",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "problem_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Problem Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "파이프라인 run 목록 (영속화된 과거 + 진행 중)",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "출제 파이프라인 실행 시작",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/delete": {
      "post": {
        "description": "body의 ids에 해당하는 run을 삭제. 진행 중 run이면 인메모리 SSE 큐도 정리.",
        "operationId": "bulk_delete_runs_api_runs_delete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bulk Delete Runs Api Runs Delete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "선택한 run 일괄 삭제",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "delete": {
        "operationId": "delete_run_api_runs__run_id__delete",
        "parameters": [
          {
            "description": "삭제할 run_id",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "삭제할 run_id",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Run Api Runs  Run Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "run 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "run 삭제",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "description": "run_id",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "run_id",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "run 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "파이프라인 run 상세 — node_states 포함 (LangSmith 토큰 lazy 보강)",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "description": "POST /api/runs가 반환한 run_id",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "POST /api/runs가 반환한 run_id",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "파이프라인 진행 SSE 스트림 (라이브 run 한정)",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/retry": {
      "post": {
        "description": "원본 run의 problem_id·target_count를 그대로 새 run으로 재실행한다. from_node 부분 재실행은 LangGraph 체크포인트가 필요해 현재는 전체 재실행으로 폴백.",
        "operationId": "retry_run_api_runs__run_id__retry_post",
        "parameters": [
          {
            "description": "재실행할 원본 run_id",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "재실행할 원본 run_id",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "원본 run 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "run 재실행 (새 run_id로 전체 재실행)",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}